add component test for Status.jsx in Status.test.jsx
